Add Edit Component
